同步本地版本

+++ Include/Timer.h
@@ -3,8 +3,8 @@
 #define RELOAD_VALUE_H (65536 - 10000) / 256 //10mS For Timer3
 #define RELOAD_VALUE_L (65536 - 100002) % 256
 
-uint16_t TimerCount;
-uint16_t KeyTime; //按键定时器,用来计时按键消抖;
+unsigned int TimerCount;
+unsigned int KeyTime; //按键定时器,用来计时按键消抖;
 
 void Timer_init(void);
